requirements.txt
asgiref==3.8.1
black==25.1.0
cfgv==3.4.0
click==8.2.1
distlib==0.3.9
django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
filelock==3.18.0
flake8==7.2.0
identify==2.6.12
isort==6.0.1
mccabe==0.7.0
mypy-extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
platformdirs==4.3.8
pre-commit==4.2.0
pycodestyle==2.13.0
pyflakes==3.3.2
pyjwt==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
ruff==0.11.13
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.31.2
